[CarCareManager] Implemented functionality to delete a vehicle. Includes: - Update of the UI in the VehicleListItem to show the options for delete, edit and select vehicle. - Implement the logic to delete the vehicle. - Show a confirmation dialog before deleting the vehicle and show a message after a successful deletion.

+++ CarCareManager/composeApp/src/commonMain/kotlin/com/echcoding/carcaremanager/data/database/dao/VehicleDao.kt
@@ -19,7 +19,7 @@ interface VehicleDao {
 
     // This function will delete a vehicle from the database
     @Query("DELETE FROM vehicles WHERE id = :id")
-    suspend fun deleteVehicleById(id: Int)
+    suspend fun deleteVehicleById(id: Long)
 
     // This function will retrieve a list of all vehicles from the database
     @Query("SELECT * FROM vehicles")
